+++ archive/old/p1_v1.go
@@ -1,4 +1,4 @@
-package main
+package archive
 
 import (
 	"bufio"
@@ -7,7 +7,7 @@ import (
 	"os"
 )
 
-func main() {
+func mainp1() {
 	file, _ := os.Open("//Users/huajiezhang/go/src/project1/data/problems.csv")
 	csvR := csv.NewReader(file)
 	scanner := bufio.NewScanner(os.Stdin)
