+++ douban_movie/main.go
@@ -1,0 +1,11 @@
+package main
+
+import (
+	"fmt"
+	"learn/test/douban_movie/douban"
+)
+
+func main() {
+	fmt.Println("hello world")
+	douban.RunAll()
+}
