--- douban_movie/test_adhoc.go
@@ -1,16 +1,0 @@
-package main
-
-import (
-	"fmt"
-	"log"
-	"net/url"
-)
-
-func main() {
-	u, err := url.Parse("https://movie.douban.com/subject/35594791/")
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	fmt.Println(u)
-}
